# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginBody {
  username: string;
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  token: string;
}

export interface AdminStats {
  totalUsers: number;
  totalPosts: number;
  totalSchools: number;
  bannedUsers: number;
  flaggedPosts: number;
  newUsersToday: number;
  newPostsToday: number;
  universities: number;
  highSchools: number;
}

export type SchoolType = (typeof SchoolType)[keyof typeof SchoolType];

export const SchoolType = {
  university: "university",
  high_school: "high_school",
} as const;

export interface School {
  id: string;
  name: string;
  type: SchoolType;
  location: string;
  county: string;
  userCount?: number;
  createdAt?: string;
}

export type CreateSchoolBodyType =
  (typeof CreateSchoolBodyType)[keyof typeof CreateSchoolBodyType];

export const CreateSchoolBodyType = {
  university: "university",
  high_school: "high_school",
} as const;

export interface CreateSchoolBody {
  name: string;
  type: CreateSchoolBodyType;
  location: string;
  county: string;
}

export type UpdateSchoolBodyType =
  (typeof UpdateSchoolBodyType)[keyof typeof UpdateSchoolBodyType];

export const UpdateSchoolBodyType = {
  university: "university",
  high_school: "high_school",
} as const;

export interface UpdateSchoolBody {
  name?: string;
  type?: UpdateSchoolBodyType;
  location?: string;
  county?: string;
}

export interface AdminUser {
  id: string;
  name: string;
  username: string;
  email: string;
  schoolName: string;
  postCount: number;
  followerCount: number;
  isBanned: boolean;
  createdAt: string;
}

export interface AdminPost {
  id: string;
  content: string;
  authorName: string;
  authorUsername: string;
  schoolName: string;
  likes: number;
  comments: number;
  isFlagged: boolean;
  createdAt: string;
}

export type ListSchoolsParams = {
  type?: ListSchoolsType;
  search?: string;
};

export type ListSchoolsType =
  (typeof ListSchoolsType)[keyof typeof ListSchoolsType];

export const ListSchoolsType = {
  university: "university",
  high_school: "high_school",
} as const;

export type ListUsersParams = {
  search?: string;
  schoolId?: string;
  banned?: boolean;
};

export type ListPostsParams = {
  search?: string;
  flagged?: boolean;
};