# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Admin login
 */
export const AdminLoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const AdminLoginResponse = zod.object({
  success: zod.boolean(),
  token: zod.string(),
});

/**
 * @summary Get platform stats
 */
export const GetAdminStatsResponse = zod.object({
  totalUsers: zod.number(),
  totalPosts: zod.number(),
  totalSchools: zod.number(),
  bannedUsers: zod.number(),
  flaggedPosts: zod.number(),
  newUsersToday: zod.number(),
  newPostsToday: zod.number(),
  universities: zod.number(),
  highSchools: zod.number(),
});

/**
 * @summary List all schools
 */
export const ListSchoolsQueryParams = zod.object({
  type: zod.enum(["university", "high_school"]).optional(),
  search: zod.coerce.string().optional(),
});

export const ListSchoolsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  type: zod.enum(["university", "high_school"]),
  location: zod.string(),
  county: zod.string(),
  userCount: zod.number().optional(),
  createdAt: zod.string().optional(),
});
export const ListSchoolsResponse = zod.array(ListSchoolsResponseItem);

/**
 * @summary Create a school
 */
export const CreateSchoolBody = zod.object({
  name: zod.string(),
  type: zod.enum(["university", "high_school"]),
  location: zod.string(),
  county: zod.string(),
});

/**
 * @summary Get a school by ID
 */
export const GetSchoolParams = zod.object({
  id: zod.coerce.string(),
});

export const GetSchoolResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  type: zod.enum(["university", "high_school"]),
  location: zod.string(),
  county: zod.string(),
  userCount: zod.number().optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Update a school
 */
export const UpdateSchoolParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateSchoolBody = zod.object({
  name: zod.string().optional(),
  type: zod.enum(["university", "high_school"]).optional(),
  location: zod.string().optional(),
  county: zod.string().optional(),
});

export const UpdateSchoolResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  type: zod.enum(["university", "high_school"]),
  location: zod.string(),
  county: zod.string(),
  userCount: zod.number().optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Delete a school
 */
export const DeleteSchoolParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List all users
 */
export const ListUsersQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  schoolId: zod.coerce.string().optional(),
  banned: zod.coerce.boolean().optional(),
});

export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  username: zod.string(),
  email: zod.string(),
  schoolName: zod.string(),
  postCount: zod.number(),
  followerCount: zod.number(),
  isBanned: zod.boolean(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Ban a user
 */
export const BanUserParams = zod.object({
  id: zod.coerce.string(),
});

export const BanUserResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  username: zod.string(),
  email: zod.string(),
  schoolName: zod.string(),
  postCount: zod.number(),
  followerCount: zod.number(),
  isBanned: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Unban a user
 */
export const UnbanUserParams = zod.object({
  id: zod.coerce.string(),
});

export const UnbanUserResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  username: zod.string(),
  email: zod.string(),
  schoolName: zod.string(),
  postCount: zod.number(),
  followerCount: zod.number(),
  isBanned: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary List all posts
 */
export const ListPostsQueryParams = zod.object({
  search: zod.coerce.string().optional(),
  flagged: zod.coerce.boolean().optional(),
});

export const ListPostsResponseItem = zod.object({
  id: zod.string(),
  content: zod.string(),
  authorName: zod.string(),
  authorUsername: zod.string(),
  schoolName: zod.string(),
  likes: zod.number(),
  comments: zod.number(),
  isFlagged: zod.boolean(),
  createdAt: zod.string(),
});
export const ListPostsResponse = zod.array(ListPostsResponseItem);

/**
 * @summary Delete a post
 */
export const DeletePostParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Flag a post for review
 */
export const FlagPostParams = zod.object({
  id: zod.coerce.string(),
});

export const FlagPostResponse = zod.object({
  id: zod.string(),
  content: zod.string(),
  authorName: zod.string(),
  authorUsername: zod.string(),
  schoolName: zod.string(),
  likes: zod.number(),
  comments: zod.number(),
  isFlagged: zod.boolean(),
  createdAt: zod.string(),
});