# lib/api-zod/src/generated/types/adminLoginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginBody {
  username: string;
  password: string;
}