# lib/api-zod/src/generated/types/adminPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminPost {
  id: string;
  content: string;
  authorName: string;
  authorUsername: string;
  schoolName: string;
  likes: number;
  comments: number;
  isFlagged: boolean;
  createdAt: string;
}