# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalPosts: number;
  totalSchools: number;
  bannedUsers: number;
  flaggedPosts: number;
  newUsersToday: number;
  newPostsToday: number;
  universities: number;
  highSchools: number;
}