# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: string;
  name: string;
  username: string;
  email: string;
  schoolName: string;
  postCount: number;
  followerCount: number;
  isBanned: boolean;
  createdAt: string;
}