# lib/api-zod/src/generated/types/createSchoolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSchoolBodyType } from "./createSchoolBodyType";

export interface CreateSchoolBody {
  name: string;
  type: CreateSchoolBodyType;
  location: string;
  county: string;
}