# lib/api-zod/src/generated/types/school.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SchoolType } from "./schoolType";

export interface School {
  id: string;
  name: string;
  type: SchoolType;
  location: string;
  county: string;
  userCount?: number;
  createdAt?: string;
}