# lib/api-zod/src/generated/types/updateSchoolBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * StudentConnect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateSchoolBodyType } from "./updateSchoolBodyType";

export interface UpdateSchoolBody {
  name?: string;
  type?: UpdateSchoolBodyType;
  location?: string;
  county?: string;
}